requirements.txt
lxml==4.9.1
python-docx==0.8.11